backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
openai>=1.23.2
httpx>=0.27.0
transformers>=4.39.3
torch>=2.2.0
torchvision>=0.17.0
huggingface-hub>=0.22.0
pydantic>=2.6.0
python-multipart>=0.0.9
requests>=2.31.0
fsspec>=2024.3.1
regex>=2023.12.25
safetensors>=0.4.0
sympy>=1.12
typing-inspection>=0.3.0
typing-extensions>=4.10.0
numpy>=1.25.0
networkx>=3.2.0
Jinja2>=3.1.2
click>=8.1.8
colorama>=0.4.6
filelock>=3.13.1
sniffio>=1.3.0
packaging>=23.2
pillow>=10.2.0
PyYAML>=6.0.1
certifi>=2025.1.31
idna>=3.7
h11>=0.14.0
distro>=1.9.0
annotated-types>=0.7.0
python-dotenv>=1.0.1